requirements.txt
feedparser==6.0.12
requests==2.32.5
openai==2.29.0
python-dotenv==1.2.2
supabase==2.28.2
openai-whisper==20250625
pytest==9.0.2
ragas
